fix(toolgen): stop double-confining spawned tools — spawn the wrapped spec plainly
spawnGeneratedTool previously called runner.spawn(...) on a spec that buildToolSpawnSpec had
already retargeted through wrapServerSpec's self-relaunch (selfSpawn("sandbox", ...)). That
launched the gateway's __nimbus-sandbox role process inside an OS sandbox, and that role process
then built its OWN SandboxRunner internally and confined the real command again — a shape with no
precedent anywhere else in the tree (every other wrapServerSpec caller, e.g.
connectors/lazy-mesh/user-mcp.ts, spawns the wrapped spec plainly).

Fix: spawnGeneratedTool now spawns the wrapServerSpec-wrapped spec directly via Bun.spawn, with no
SandboxRunner parameter at all — single confinement layer, performed by the __nimbus-sandbox role
process the spec already re-launches into. Removed the now-unused policyFromManifest/SandboxRunner
imports.

Replaced the Node ChildProcess-based ToolChildIo adapter with an exported ioFromSpawnedChild
(Bun.Subprocess-based, generic over the stderr mode), used by both spawnGeneratedTool and the
runtime round-trip test — so the test drives the exact same adapter code production spawns
through, not a second reimplementation.

The round-trip test still spawns its own throwaway child via a raw Bun.spawn (bypassing
buildToolSpawnSpec/wrapServerSpec entirely), since going through the wrapped spec would now
actually invoke real OS sandbox confinement from a unit test — out of scope here per Task 18.

Coverage assertions, CANONICAL, and the egress-boot-marker.test.ts fix from the prior commit are
unchanged.
